+++ src/parsing/java/parser.rs
@@ -90,9 +90,9 @@ impl JavaParser {
         let end = node.end_position();
         Range {
             start_line: start.row as u32,
-            start_column: start.column as u16,
+            start_column: start.column as u32,
             end_line: end.row as u32,
-            end_column: end.column as u16,
+            end_column: end.column as u32,
         }
     }
