+++ src/parsing/swift/parser.rs
@@ -124,9 +124,9 @@ impl SwiftParser {
         let end = node.end_position();
         Range {
             start_line: start.row as u32,
-            start_column: start.column as u16,
+            start_column: start.column as u32,
             end_line: end.row as u32,
-            end_column: end.column as u16,
+            end_column: end.column as u32,
         }
     }
 
@@ -1671,9 +1671,9 @@ impl SwiftParser {
     fn binding_range(node: Node) -> Range {
         Range::new(
             node.start_position().row as u32,
-            node.start_position().column as u16,
+            node.start_position().column as u32,
             node.end_position().row as u32,
-            node.end_position().column as u16,
+            node.end_position().column as u32,
         )
     }
